Remove bundled expect.el
The library expect.el was written by Lars Magne Ingebrigtsen and
nowadays he maintains it in the repository of the `cddb' package.
That is an interface to the "CD database", which isn't really related
to `mailcrypt'.  Nevertheless I think it is better to depend on the
library from the `cddb' repository instead of bundling an outdated
version.